Item1776: don't check validation key when running from the command-line
git-svn-id: http://svn.foswiki.org/trunk@4603 0b4bb1d4-4e5a-0410-9cc4-b2b747904278

+++ core/lib/Foswiki/UI.pm
@@ -519,6 +519,9 @@ sub checkValidationKey {
     # If validation is disabled, do nothing
     return if ( $Foswiki::cfg{Validation}{Method} eq 'none' );
 
+    # No point in command-line mode
+    return if $session->inContext('command_line');
+
     # Check the nonce before we do anything else
     my $nonce = $session->{request}->param('validation_key');
     $session->{request}->delete('validation_key');
